Final Application/requirements.txt
Flask==1.0.2
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-cloud-automl==2.2.0
grpcio==1.37.0
gunicorn==19.9.0
spacy[lookups]==2.3.0
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-2.3.0/fr_core_news_sm-2.3.0.tar.gz#egg=fr_core_news_sm==2.3.0
deep_translator==1.4.4